fix: Address PR review comments for training and adapters
- Add documentation explaining &mut self requirement in training_step
- Retrieve pad token ID from tokenizer instead of hard-coding 0
- Mask padding tokens with -100 in labels for proper loss computation
- Implement custom loss computation that ignores padding tokens
- Change optimizer step to error instead of silently skipping
- Change checkpoint save to error if optimizer not initialized
- Remove redundant type casts in adapter config conversions
- Add learning_rate assertion to checkpoint save/load test
- Initialize optimizer in test to match production usage pattern

+++ src/adapters/mod.rs
@@ -89,8 +89,8 @@ impl AdapterWrapper {
     pub fn to_peft_lora_config(settings: &LoraSettings) -> PeftLoraConfig {
         PeftLoraConfig {
             r: settings.r,
-            alpha: settings.alpha as usize,
-            dropout: settings.dropout as f64,
+            alpha: settings.alpha,
+            dropout: settings.dropout,
             target_modules: settings.target_modules.clone(),
             ..Default::default()
         }
@@ -157,7 +157,7 @@ impl From<&LoraSettings> for AdapterApplicationConfig {
         Self {
             target_modules: settings.target_modules.clone(),
             r: settings.r,
-            alpha: settings.alpha as usize,
+            alpha: settings.alpha,
             dropout: settings.dropout as f32,
         }
     }
